IPA ss7_link: Construct fake mtp3_routing_label{}
In SCCP-over-IPA links, we don't have MTP3 or something else that can
transport a routing label.  It's a true point-to-point connection and
it cannot do MTP-based routing.  In order to fit this into our existing
SS7 link / linkset code, we locally construct a fake mtp3 routing label
with the remote and local point code of the linkset.